Update sample import script to use a valid value for MaxSizeBytes (#340)
The value previously in use (262144000, which is 250MB) is not a valid size for an S3 database. Trying to create a database of that size will throw an error like `The edition 'Standard' does not support the database data max size '262144000'.`

I am not clear whether 250MB was *ever* a valid size for S3, but it certainly is not now. It used to be the case that the Import/Export service would silently swallow this error and just create a database of the max size for the SLO. However, that (buggy) behavior was fixed around November 2020 to respect the parameter.